notification 읽기처리 api,로직추가, 읽기 이후 navigate,targetDeleted 처리 로직 추가

+++ domain/src/main/java/com/androiddev/domain/model/NotificationItem.kt
@@ -1,9 +1,9 @@
 package com.androiddev.domain.model
 
 data class NotificationExtra(
-    val postId: Long? = null,
-    val commentId: Long? = null,
-    val followerId: Long? = null
+    val postId: Int? = null,
+    val commentId: Int? = null,
+    val followerId: Int? = null
 )
 
 data class NotificationItem(
